+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+* finagle-thrift: Add `c.t.finagle.thrift.MethodMetadata` which provides a `LocalContext` Key
+  for setting information about the current Thrift method and an accessor for retrieving 
+  the currently set value. ``PHAB_ID=D241295``
+
 18.11.0
 -------